requirements-dev.txt
# Development dependencies for testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
bleak>=0.21.0
pillow>=10.0.0
